refactor : 피드백 반영 및 패키지 이동
