+++ src/test/java/calculator/domain/StringAddCalculatorTest.java
@@ -1,5 +1,6 @@
-package calculator;
+package calculator.domain;
 
+import calculator.domain.StringAddCalculator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
